% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Genseq}
\alias{Genseq}
\title{Translate DNA or RNA to amino acide}
\usage{
Genseq(x)
}
\arguments{
\item{x}{a text string containing a DNA or RNA sequence}
}
\value{
A text string that represents the translated amino acid sequence.
}
\description{
Translate DNA or RNA to amino acide
}
\examples{
Translate(x)
x <- "ATGTCCTAG"
Translate(x)
}
